feat: add support for coverage
In order to support check of the coverage of unit tests, `karma-coverage` uses `istanbul` package. This package injects some custom code in the .js files before executing the tests. After that the tests are executed with the new files and coverage information is collected.
This does not work with NativeScript Unit-Test Runner, as the files on device do not contain the injected code. However, we can ask karma to serve them (similar to what we've already implemented for the testing framework) and after that we can overwrite the content in the application itself.
Then the tests are executed and karma creates coverage information.

NOTE: This is just POC for one way to implement the feature. It will not work with platform specific files as when the device asks karma to serve a file, it will fail as it will not find such file in the local app dir.